Add signup  and login pages
