+++ pages/signup/index.tsx
@@ -1,0 +1,246 @@
+'use client'
+import React, { useState, useRef } from 'react'
+import { FidgetSpinner } from 'react-loader-spinner';
+import style from '../../../../assets/stylesheets/signup.module.css';
+import { useDispatch, useSelector, Provider } from 'react-redux';
+import { registerAdmin } from '../../src/app/redux/actions/authSlice';
+import { ToastContainer, toast } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
+import store from '../../src/app/redux/store';
+import { useRouter } from 'next/router';
+import '../../src/app/globals.css';
+import Link from 'next/link';
+
+const SignUp = () => {
+
+    const loading = useSelector((state) => state.auth.loading);
+    const AdminInfo = useSelector((state) => state.auth.admin);
+
+    const router = useRouter();
+
+    const dispatch = useDispatch();
+    const formRef = useRef(null);
+    const [admin, setAdmin] = useState({
+        first_name: '',
+        last_name: '',
+        email: '',
+        password: ''
+    })
+
+    const [password, setPassword] = useState('')
+    const [isPasswordMatched, setIsPasswordMatched] = useState(false);
+    const [isAdminCode, setIsAdminCode] = useState(false);
+    const [adminPasscode, setAdminPasscode] = useState('');
+
+    const handleAdminCodeVerification = (code) => {
+
+        const span = document.getElementById('admin-code-error');
+
+        setAdminPasscode(prevPasscode => {
+            const newPasscode = Number(code);
+            const ADMINCODE = 50326;
+
+            if (newPasscode === ADMINCODE) {
+                setIsAdminCode(true)
+                span!.innerText = 'yeah ✔';
+            } else {
+                setIsAdminCode(false)
+                span!.innerText = 'Invalid Admin code';
+            }
+            return newPasscode;
+        });
+    }
+
+    const validatePasswordMatch = (password, confirmPassword) => {
+        const passwordInfo = document.getElementById('password-error');
+
+        if (confirmPassword === '') {
+            passwordInfo!.innerText = '';
+            setAdmin(
+                { ...admin, password: '' }
+            )
+        } else if (password === confirmPassword) {
+            setIsPasswordMatched(true);
+            setAdmin({ ...admin, password: password });
+            passwordInfo!.innerText = 'Passwords Matched ✔';
+        } else if (password === '' || password !== confirmPassword) {
+            setIsPasswordMatched(false);
+            passwordInfo!.innerText = 'Passwords do not match..';
+            setAdmin(
+                { ...admin, password: '' }
+            )
+        } else {
+
+        }
+    }
+
+    const validateEmail = (email) => {
+
+        const emailRegex = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
+
+        const span = document.getElementById('email-error');
+
+        const isEmailValid = emailRegex.test(email);
+
+        if (isEmailValid) {
+            setAdmin({ ...admin, email: email })
+            span!.innerText = '';
+        } else if (email === '') {
+            span!.innerText = '';
+        } else if (!isEmailValid) {
+            span!.innerText = 'Invalid email';
+        }
+    }
+
+    const handleEmail = (e) => {
+        validateEmail(e.target.value);
+    }
+
+    const handleSubmit = async (e) => {
+        e.preventDefault();
+        console.log(admin);
+
+        const allInputHasValue = Object.entries(admin).every(([key, value]) => {
+            const message = document.getElementById(`${key}-error`);
+            if (value === '') {
+                message!.innerText = `Please Enter ${key}`;
+                return false
+            }
+            message!.innerText = '';
+            return true;
+        });
+
+        if (allInputHasValue) {
+            try {
+
+                const response = await dispatch(registerAdmin(admin))
+                console.log(response.payload.status.errors);
+                if (!response.payload.status.errors) {
+                    toast.success('Sign up successful');
+                    router.push('/');
+                } else {
+                    toast.error(response.payload.status.errors[0])
+                }
+
+
+            } catch (error) {
+                console.log('Error logging in:', error);
+                toast.error('An error occurred while logging in. Please try again later.');
+            }
+
+        }
+    }
+
+
+        return (
+            <>
+
+                {loading && (
+                    <div className='flex flex-col absolute w-full h-full bg-white justify-center items-center'>
+                        <FidgetSpinner />
+                        <h4 className='font-bold'>Signing up</h4>
+                    </div>
+                )}
+                <div className={`text`}>
+
+                    <ToastContainer />
+                    <form className={` flex opacity-100 flex-col w-[85%] md:w-full space-y-10 m-auto`} ref={formRef} onSubmit={(e) => handleSubmit(e)}>
+
+                        <h1 className={` text-center text-3xl font-bold `}>Sign Up</h1>
+                        <label htmlFor='first_name'>
+                            <input
+                                className='border rounded-md p-2 w-full'
+                                id='first_name'
+                                type="text"
+                                name="first_name"
+                                onChange={(e) => setAdmin({ ...admin, first_name: e.target.value })}
+                                placeholder='First Name..'
+                            />
+                            <small className='text-rose-500' id='first_name-error'></small>
+                        </label>
+
+                        <label htmlFor='last_name'>
+                            <input
+                                className='border rounded-md p-2 w-full'
+                                id='last_name'
+                                type="text"
+                                name="last_name"
+                                onChange={(e) => setAdmin({ ...admin, last_name: e.target.value })}
+                                placeholder='Last Name..'
+                            />
+                            <small className='text-rose-500' id='last_name-error'></small>
+                        </label>
+
+                        <label htmlFor='email'>
+                            <input
+                                className='border rounded-md p-2 w-full'
+                                id='email'
+                                type="email"
+                                name="email_input"
+                                onChange={(e) => {
+                                    setAdmin({ ...admin, email: e.target.value })
+                                    handleEmail(e)
+                                }}
+                                placeholder='Email..'
+                            />
+                            <small className='text-rose-500' id='email-error'></small>
+                        </label>
+
+                        <label htmlFor='admin_verification'>
+                            <input
+                                className='border rounded-md p-2 w-full'
+                                type="text"
+                                id='admin_verification'
+                                name="admin_verification"
+                                onChange={(e) => handleAdminCodeVerification(e.target.value)}
+                                placeholder='Enter Admin authentication code...'
+                            />
+                            <small className={isAdminCode ? 'text-emerald-400' : 'text-rose-500'} id='admin-code-error'></small>
+                        </label>
+
+                        <label htmlFor='password'>
+                            <input
+                                className='border rounded-md p-2 w-full'
+                                type="password"
+                                name="password_input"
+                                id='password'
+                                onChange={(e) => setPassword(e.target.value)}
+                                placeholder='Password'
+                            />
+                        </label>
+
+                        <label htmlFor='password_confirmation'>
+                            <input
+                                className='border rounded-md p-2 w-full'
+                                type="password"
+                                name="password_confirmation"
+                                id='password_confirmation'
+                                onChange={(e) => validatePasswordMatch(password, e.target.value)}
+                                placeholder='Confirm Password'
+                            />
+                            <small className={isPasswordMatched ? 'text-emerald-400' : 'text-rose-500'} id='password-error'></small>
+                        </label>
+
+                        <input
+                            className='border w-1/3 rounded-lg p-2 m-auto hover:bg-slate-800 hover:text-white font-bold text-lg'
+                            type="submit"
+                            value="Submit"
+                        />
+                    </form>
+
+                    <p className="text-center absolute w-full mt-3">Already have an account? <Link href="/login">Login</Link></p>
+
+                </div>
+            </>
+        )
+
+    }
+
+    const SignupPage = () => (
+
+        <Provider store={store}>
+            <SignUp />
+        </Provider>
+    );
+
+    export default SignupPage;
