+++ src/app/components/Navbar.tsx
@@ -7,11 +7,21 @@ import Link from 'next/link';
 import Image from 'next/image';
 import logo from '@/assets/logo2.png'
 import { motion } from 'framer-motion';
+import Cookies from 'js-cookie';
+import { ToastContainer, toast } from 'react-toastify';
+import { useDispatch, useSelector} from 'react-redux';
+import { useRouter } from 'next/navigation'
+import { logoutAdmin } from '../../../src/app/redux/actions/authSlice';
 
 const Navbar = () => {
     const [menu, setMenu] = useState(false);
     const [prevScrollPos, setPrevScrollPos] = useState(0);
     const [visible, setVisible] = useState(true);
+
+    const dispatch = useDispatch();
+
+    const loggedIn = Cookies.get('admin_token');
+    const router = useRouter();
 
     useEffect(() => {
         const handleScroll = () => {
@@ -24,6 +34,15 @@ const Navbar = () => {
 
         return () => window.removeEventListener('scroll', handleScroll);
     }, [prevScrollPos, visible]);
+
+    const handleLogout = (e) => {
+        e.stopPropagation();
+        dispatch(logoutAdmin());
+        toast.success('Logged out successful');
+        setTimeout(() => {
+          router.push('/login');
+        }, 1000);
+      };
 
     const navLinks = [
         { title: "Home", link: "/" },
@@ -41,6 +60,7 @@ const Navbar = () => {
                 { 'transform': !visible, '-translate-y-full': !visible, 'shadow-lg': !visible }
             )}
         >
+            <ToastContainer />
             <section className='flex top-0 w-full justify-between'>
                 <MdMenu className="text-3xl cursor-pointer lg:hidden" onClick={() => (setMenu(true), console.log(menu))} />
                 <div className='flex px-3'>
@@ -54,6 +74,23 @@ const Navbar = () => {
                     <button className='hidden md:flex p-3 rounded-full  bg-blue-600 lg:block text-1xl text-grey-20 hover:bg-blue-400 hover:text-white transition-all  duration-500 ease-in-out z-40 transform  border border-black'>
                         Upgrade
                     </button>
+                    {loggedIn ? (
+    <button
+        name='button'
+        type='button'
+        onClick={handleLogout}
+        className='hidden p-3 rounded lg:block text-2xl text-grey-20 hover:bg-blue-400 hover:text-white transition-all delay-0 duration-300 ease-out z-40 transform'
+    >
+        Sign Out
+    </button>
+) : (
+    <Link
+        href='/login'
+        className='hidden p-3 rounded lg:block text-2xl text-grey-20 hover:bg-blue-400 hover:text-white transition-all delay-0 duration-300 ease-out z-40 transform'
+    >
+        Log In
+    </Link>
+)}
                 </div>
             </section>
             {/* menu */}
@@ -66,6 +103,24 @@ const Navbar = () => {
                     {navLinks.map((d, i) => (
                         <Link key={i} href={d.link} className='text-2xl font-bold hover:bg-black rounded transition-all duration-75 w-fit p-4'>{d.title}</Link>
                     ))}
+
+{loggedIn ? (
+    <button
+        name='button'
+        type='button'
+        onClick={handleLogout}
+        className='text-2xl font-bold hover:bg-black rounded transition-all duration-75 w-fit p-4'
+    >
+        Sign Out
+    </button>
+) : (
+    <Link
+        href='/login'
+        className='text-2xl font-bold hover:bg-black rounded transition-all duration-75 w-fit p-4'
+    >
+        Log In
+    </Link>
+)}
                 </section>
             </section>
         </motion.nav>
