+++ src/app/redux/actions/authSlice.tsx
@@ -1,54 +1,71 @@
 import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
 import axios from 'axios';
 
-const BASE_URL = 'http://127.0.0.1:3000/users';
+import Cookies from 'js-cookie';
+
+const BASE_URL = 'http://127.0.0.1:3000/admins';
 
 
-export const loginUser = createAsyncThunk(
-    'auth/loginUser',
-    async (user) => {
-        const response = await axios.post(`${BASE_URL}/login`, user)
+export const loginAdmin = createAsyncThunk(
+    'auth/loginAdmin',
+    async (admin) => {
+        const response = await axios.post(`${BASE_URL}/sign_in`, admin)
         return response.data
     }
 )
 
+export const logoutAdmin = createAsyncThunk(
+    'auth/logoutAdmin',
+    async () => {
+        const response = await fetch(`${BASE_URL}/sign_out`, {
+            method: 'DELETE',
+            headers: {
+                'Content-Type': 'application/json',
+                'Authorization': Cookies.get('admin_token')
+            }
+        })
+
+        Cookies.remove('admin_token')
+        const data = await response.json();
+        console.log(data)
+        return data
+    }
+)
+
 // Define a type or interface for the user object
-interface User {
+interface Admin {
     email: string;
     password: string;
     first_name: string;
     last_name: string;
-    role: string;
 }
 
 // Create an async thunk for registering a user
-export const registerUser = createAsyncThunk(
-    'auth/registerUser',
-    async (user: User) => { // Type the user parameter
-        const userObject = {
-            user: {
-                email: user.email,
-                password: user.password,
-                first_name: user.first_name,
-                last_name: user.last_name,
-                role: user.role
+export const registerAdmin = createAsyncThunk(
+    'auth/registerAdmin',
+    async (admin: Admin) => { // Type the user parameter
+        const adminObject = {
+            admin: {
+                email: admin.email,
+                password: admin.password,
+                first_name: admin.first_name,
+                last_name: admin.last_name,
             }
         };
-
-        // Assume BASE_URL is defined somewhere in your code
-        const BASE_URL = 'your_api_base_url';
 
         const response = await fetch(BASE_URL, {
             method: 'POST',
             headers: {
                 'Content-Type': 'application/json'
             },
-            body: JSON.stringify(userObject)
+            body: JSON.stringify(adminObject)
         });
 
         if (!response.ok) {
-            throw new Error('Failed to register user');
+            throw new Error('Failed to register Admin');
         }
+        console.log(response.headers.get('Authorization'))
+         await Cookies.set('admin_token', response.headers.get('Authorization'), { expires: 1 })
 
         const data = await response.json();
         return data;
@@ -56,7 +73,7 @@ export const registerUser = createAsyncThunk(
 );
 
 const initialState = {
-    user: null,
+    admin: null,
     status: 'idle',
     error: null,
     loading: false
@@ -69,31 +86,45 @@ export const authSlice = createSlice({
     },
     extraReducers: (builder) => {
         builder
-            .addCase(loginUser.pending, (state) => {
+            .addCase(loginAdmin.pending, (state) => {
                 state.loading = true
                 state.status = 'loading'
             })
-            .addCase(loginUser.fulfilled, (state, action) => {
+            .addCase(loginAdmin.fulfilled, (state, action) => {
                 state.loading = false
                 state.status = 'succeeded'
                 state.user = action.payload
             })
-            .addCase(loginUser.rejected, (state, action) => {
+            .addCase(loginAdmin.rejected, (state, action) => {
                 state.loading = false
                 state.status = 'failed'
-                state!.error = action.error.message
+                state.error = action.error.message
             })
-            .addCase(registerUser.pending, (state) => {
+            .addCase(registerAdmin.pending, (state) => {
                 state.loading = true
                 state.status = 'loading'
             })
-            .addCase(registerUser.fulfilled, (state, action) => {
+            .addCase(registerAdmin.fulfilled, (state, action) => {
                 state.loading = false
                 state.status = 'succeeded'
                 state.user = action.payload
                 console.log(state)
             })
-            .addCase(registerUser.rejected, (state, action) => {
+            .addCase(registerAdmin.rejected, (state, action) => {
+                state.loading = false
+                state.status = 'failed'
+                state.error = action.error.message
+            })
+            .addCase(logoutAdmin.pending, (state) => {
+                state.loading = true
+                state.status = 'loading'
+            })
+            .addCase(logoutAdmin.fulfilled, (state, action) => {
+                state.loading = false
+                state.status = action.payload
+                state.user = null
+            })
+            .addCase(logoutAdmin.rejected, (state, action) => {
                 state.loading = false
                 state.status = 'failed'
                 state.error = action.error.message
