+++ src/assets/stylesheets/login.module.css
@@ -9,7 +9,7 @@
 .form_title {
   font-size: 2rem;
   font-weight: 700;
-  color: #f5f5f5;
+  color: #c07c7c;
   text-align: center;
   margin-bottom: 1rem;
 }
